+++ test/unit/fuzz-tools.test.js
@@ -411,6 +411,7 @@ import * as bellNecV431 from '../../lib/bell-nec-v431.js';
 import * as badenWalkerV432 from '../../lib/baden-walker-v432.js';
 import * as modicChangesV433 from '../../lib/modic-changes-v433.js';
 import * as pfirrmannDiscV434 from '../../lib/pfirrmann-disc-v434.js';
+import * as vanHerickV435 from '../../lib/van-herick-v435.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -767,6 +768,7 @@ const MODULES = {
   'baden-walker-v432.js': badenWalkerV432,
   'modic-changes-v433.js': modicChangesV433,
   'pfirrmann-disc-v434.js': pfirrmannDiscV434,
+  'van-herick-v435.js': vanHerickV435,
   'tb-testing.js': tbTesting,
 };